amqpspawn: option to limit the number of messages to receive
Adds an option (--number/-n) to limit the number of messages an instance
of the program will accept from the server. Also adjusts the prefetch
limit on the connection to 10 by default, changing to the user specified
value if supplied. WARNING: Using --number option with --no-ack is VERY
likely to cause data loss when connecting to a server with queued
(persistent) messages. If you care about this, do NOT use the --no-ack
option.