attempt to allow arguments to external program
I can't program C and I can't get the damn thing to work so this is the best I
can get. It doesn't work. For some reason execl cuts off the first character of
program_args when passing it to the external program. For example, if my
external program should be:
	foobar -D -m foo.txt routingkey tmpfile
The executed program that execl spawns is:
	foobar D -m foo.txt routingkey tmpfile